+++ src/ONTraC/bin/ONTraC_GNN.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 
+import os
 import sys
 from os import path
 from typing import Callable, Optional
@@ -43,6 +44,9 @@ def main() -> None:
 
     # load parameters
     options = load_parameters(opt_validate_func=opt_GNN_validate, prepare_optparser_func=prepare_GNN_optparser)
+    options.dataset = f'{options.preprocessing_dir}/original_data.csv'
+    if not os.path.exists(options.dataset):
+        raise FileNotFoundError(f"Dataset file not found: {options.dataset}. You may need to run createDataSet first or copy original dataset file into {options.preprocessing_dir} directory with the name 'original_data.csv'.")
 
     # load meta data
     options.meta_input = path.join(options.preprocessing_dir, 'meta_data.csv')