+++ src/ONTraC/bin/ONTraC_analysis.py
@@ -5,11 +5,13 @@ from optparse import OptionGroup, OptionParser, Values
 from ..analysis.cell_type import cell_type_visualization
 from ..analysis.data import AnaData
 from ..analysis.niche_cluster import niche_cluster_visualization
-from ..analysis.niche_net import clustering_visualization, embedding_adjust_visualization
+from ..analysis.niche_net import (clustering_visualization,
+                                  embedding_adjust_visualization)
 from ..analysis.spatial import spatial_visualization
 from ..analysis.train_loss import train_loss_visualiztion
 from ..log import *
-from ..optparser._IO import add_IO_options_group, validate_io_options, write_io_options_memo
+from ..optparser._IO import (add_IO_options_group, validate_io_options,
+                             write_io_options_memo)
 from ..utils import *
 from ..version import __version__
 
@@ -55,32 +57,8 @@ def analysis_pipeline(options: Values) -> None:
 
     info('--------------- Analysis pipeline end --------------- ')
 
+
 # TODO: move to optparser
-def add_suppress_group(optparser: OptionParser) -> None:
-    group = OptionGroup(optparser, 'Suppress options')
-    group.add_option('--suppress-cell-type-composition',
-                     dest='suppress_cell_type_composition',
-                     action='store_true',
-                     default=False,
-                     help='Suppress the cell type composition visualization.')
-    group.add_option('--suppress-niche-cluster-loadings',
-                     dest='suppress_niche_cluster_loadings',
-                     action='store_true',
-                     default=False,
-                     help='Suppress the niche cluster loadings visualization.')
-    group.add_option('--suppress-cell-type',
-                     dest='suppress_cell_type',
-                     action='store_true',
-                     default=False,
-                     help='Suppress the cell type visualization.')
-    group.add_option('--suppress-niche-trajectory',
-                        dest='suppress_niche_trajectory',
-                        action='store_true',
-                        default=False,
-                        help='Suppress the niche trajectory related visualization.')
-    optparser.add_option_group(group)
-
-
 def add_embedding_adjust_group(optparser: OptionParser) -> None:
     group = OptionGroup(optparser, 'Embedding adjust options')
     group.add_option(
@@ -102,12 +80,55 @@ def add_embedding_adjust_group(optparser: OptionParser) -> None:
     optparser.add_option_group(group)
 
 
+def add_suppress_group(optparser: OptionParser) -> None:
+    group = OptionGroup(optparser, 'Suppress options')
+    group.add_option('--suppress-cell-type-composition',
+                     dest='suppress_cell_type_composition',
+                     action='store_true',
+                     default=False,
+                     help='Suppress the cell type composition visualization.')
+    group.add_option('--suppress-niche-cluster-loadings',
+                     dest='suppress_niche_cluster_loadings',
+                     action='store_true',
+                     default=False,
+                     help='Suppress the niche cluster loadings visualization.')
+    group.add_option('--suppress-cell-type',
+                     dest='suppress_cell_type',
+                     action='store_true',
+                     default=False,
+                     help='Suppress the cell type visualization.')
+    group.add_option('--suppress-niche-trajectory',
+                     dest='suppress_niche_trajectory',
+                     action='store_true',
+                     default=False,
+                     help='Suppress the niche trajectory related visualization.')
+    optparser.add_option_group(group)
+
+
+def add_visualization_group(optparser: OptionParser) -> None:
+    group = OptionGroup(optparser, 'Visualization options')
+    group.add_option('-s',
+                     '--sample',
+                     dest='sample',
+                     action='store_true',
+                     default=False,
+                     help='Plot each sample separately.')
+    group.add_option('--scale-factor',
+                     dest='scale_factor',
+                     type='float',
+                     default=1.0,
+                     help='Scale factor control the size of spatial-based plots.')
+    optparser.add_option_group(group)
+
+
 def prepare_optparser() -> OptionParser:
     """Prepare optparser object. New options will be added in this function first.
 
     Ret: OptParser object.
     """
-    usage = "USAGE: %prog <--preprocessing-dir PREPROCESSING_DIR> <--GNN-dir GNN_DIR> <--NTScore-dir NTSCORE_DIR> <-o OUTPUT_DIR> [-l LOG_FILE] [-r REVERSE]"
+    usage = """USAGE: %prog <--preprocessing-dir PREPROCESSING_DIR> <--GNN-dir GNN_DIR>
+    <--NTScore-dir NTSCORE_DIR> <-o OUTPUT_DIR> [-l LOG_FILE] [-r REVERSE] [-s SAMPLE] [--scale-factor SCALE_FACTOR]
+    [--suppress-cell-type-composition] [--suppress-niche-cluster-loadings] [--suppress-niche-trajectory]"""
     description = "Analysis the results of ONTraC."
     optparser = OptionParser(version=f'%prog {__version__}',
                              usage=usage,
@@ -123,14 +144,9 @@ def prepare_optparser() -> OptionParser:
                          action='store_true',
                          default=False,
                          help='Reverse the NT score.')
-    optparser.add_option('-s',
-                         '--sample',
-                         dest='sample',
-                         action='store_true',
-                         default=False,
-                         help='Plot each sample separately.')
     add_embedding_adjust_group(optparser)
     add_suppress_group(optparser)
+    add_visualization_group(optparser)
     return optparser
 
 
@@ -153,7 +169,7 @@ def opt_validate(optparser: OptionParser) -> Values:
     else:
         warning(f'Output directory already exists: {options.output}, will overwrite it.')
 
-    if not os.path.exists(options.log):
+    if options.log is not None and not os.path.exists(options.log):
         error(f'Log file not found: {options.log}')
         sys.exit(1)
 
@@ -167,13 +183,16 @@ def opt_validate(optparser: OptionParser) -> Values:
     # print parameters to stdout
     write_io_options_memo(options, IO_OPTIONS)
     info(f'Output directory: {options.output}')
-    info(f'Log file: {options.log}')
-    info('       -------- ONTraC running options --------        ')
+    if options.log is not None:
+        info(f'Log file: {options.log}')
+    info('  -------- ONTraC running options needed here -------- ')
     info(f'Reverse: {options.reverse}')
-    info(f'Sample: {options.sample}')
     info(f'Embedding adjust: {options.embedding_adjust}')
     if options.embedding_adjust:
         info(f'Sigma: {options.sigma}')
+    info(' ---------------- Visualization params --------------- ')
+    info(f'Sample: {options.sample}')
+    info(f'Scale factor: {options.scale_factor}')
     info('         -------- Suppression options --------         ')
     if hasattr(options, 'suppress_cell_type_composition'):
         info(f'Suppress cell type composition: {options.suppress_cell_type_composition}')