+++ tests/test_niche_trajectory.py
@@ -1,18 +1,9 @@
-from optparse import Values
+from unittest.mock import Mock
 
 import numpy as np
-import pytest
 
 from ONTraC.niche_trajectory._niche_trajectory import (
     get_niche_NTScore, get_niche_trajectory_path, trajectory_path_to_NC_score)
-
-
-@pytest.fixture
-def options() -> Values:
-    # Create an options object for testing
-    _options = Values()
-    _options.equal_space = True  # TODO: test for equal_space=False
-    return _options
 
 
 def test_get_niche_trajectory_path() -> None:
@@ -25,15 +16,25 @@ def test_get_niche_trajectory_path() -> None:
     # Test case: Adjacency matrix with 5 nodes
     adj_matrix = np.array([[0, 1, 0, 1, 0], [1, 0, 1, 0, 1], [0, 1, 0, 1, 0], [1, 0, 1, 0, 1], [0, 1, 0, 1, 0]])
     expected_path = [0, 1, 2, 3, 4]
-    assert get_niche_trajectory_path(adj_matrix) == expected_path
+
+    options = Mock()
+
+    options.trajectory_construct = "BF"
+    assert get_niche_trajectory_path(options=options, niche_adj_matrix=adj_matrix) == expected_path
+
+    options.trajectory_construct = "TSP"
+    assert get_niche_trajectory_path(options=options, niche_adj_matrix=adj_matrix) == expected_path
 
 
-def test_trajectory_path_to_NC_score(options: Values):
+def test_trajectory_path_to_NC_score():
     """
     Test the function trajectory_path_to_NC_score.
     :param options: Values, options.
     :return: None.
     """
+
+    options = Mock()
+    options.equal_space = True
 
     # Test case: Niche trajectory path with 6 clusters
     niche_trajectory_path = [1, 2, 3, 4, 5, 0]
@@ -45,18 +46,21 @@ def test_trajectory_path_to_NC_score(options: Values):
     assert np.allclose(gen_NT_score, expected_NT_score)
 
 
-def test_get_niche_NTScore(options: Values):
+def test_get_niche_NTScore():
     """
     Test the function get_niche_NTScore.
     :param options: Values, options.
     :return: None.
     """
-    
+
     # Test case: Niche cluster loading with 3 clusters and adjacency matrix with 3 nodes
     niche_cluster_loading = np.array([[0.1, 0.2, 0.7], [0.4, 0.5, 0.1], [0.6, 0.1, 0.3]])  # #niche x #niche_cluster
     niche_adj_matrix = np.array([[0, 1, 0], [1, 0, 1], [0, 1, 0]])
     expected_niche_cluster_score = np.array([0, 0.5, 1])
     expected_niche_level_NTScore = np.array([0.8, 0.35, 0.35])
+
+    options = Mock()
+    options.trajectory_construct = "BF"
 
     niche_cluster_score, niche_level_NTScore = get_niche_NTScore(options=options,
                                                                  niche_cluster_loading=niche_cluster_loading,